requirements.txt
appdirs==1.4.4
asgiref==3.3.1
beautifulsoup4==4.9.3
beautifultable==0.8.0
bootstrap4==0.1.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==8.0.0
coverage==5.5
cryptography==3.4.7
cssselect==1.1.0
DateTime==4.3
defusedxml==0.7.1
distlib==0.3.1
Django==3.1.4
django-adminlte-3==0.1.6
django-adminlte3==0.1.6
django-appconf==1.0.4
django-bootstrap4==3.0.0
django-braces==1.14.0
django-comments-dab==1.6.7
django-comments-xtd==2.9.0
django-contrib-comments==2.1.0
django-crispy-forms==1.11.2
django-db-comments==0.2.2
django-filter==2.4.0
django-gravatar2==1.4.4
django-model-utils==4.1.1
django-six==1.0.4
django-star-ratings==0.9.1
django-static-jquery==2.1.4
django-utils-six==2.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
docutils==0.17.1
fandogh-cli==1.26.3
fanstatic==1.1
filelock==3.0.12
flake8==3.9.1
idna==2.10
js.jquery==3.2.1
lxml==4.6.3
mccabe==0.6.1
novas==3.1.1.5
oauthlib==3.1.0
packaging==20.9
Pillow==8.0.1
pinax-comments==2.0.0
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.1.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-decouple==3.4
python3-openid==3.2.0
pytz==2020.4
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
retrying==1.3.3
shutilwhich==1.1.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.4.1
swapper==1.1.2.post1
toml==0.10.2
tox==3.23.0
tzlocal==2.1
urllib3==1.26.4
virtualenv==20.4.4
WebOb==1.8.7
websocket-client==1.0.0
zope.interface==5.4.0
